Move as_nodegroup()
